==> 6_1_make_faceted_barcharts_plots_signif_drivers.R <==
#This script creates Fig 4B,C,D,S3,S8C,D,E,F
#Input: .csv files output by script 6_0
#Output: barcharts counting absolute correlations 
        #barcharts showing the two types of correlations: increase in sensitivity with increase in activity/expression
                                                         #increase in sensitivity with deacrease in activity/expression

library(stringr)
library(magrittr)
library(tidyverse)
library(Cairo)

#select regulons
regulons <- "aracne" #OR
regulons <- "grndb" #OR
regulons <- "dorothea" #OR
regulons <- "aracne_ccle" #OR
regulons <- "grndb_ccle"

#select cancer list analysis
cancer_list = c("aml", "blca", "brca", "coad",
                "gbm", "hnsc", "kirc", "luad",
                "paad", "stad") #OR

cancer_list = c("brca","coad", "luad", "paad") #for aracne_ccle and grndb_ccle #OR

cancer_list = c("pancancer") #only works with dorothea regulons


#If the pancancer analysis is chosen - choose a threshold for essential/non-essential genes:
#1 percent (i.e., 9 cell lines) in which we are looking for a gene to be essential/non-essential in order to test it
#10 percent (i.e., 97 cell lines) in which we are looking for a gene to be essential/non-essential in order to test it
threshold_cell_lines <- "1percent"
threshold_cell_lines <- "5percent"
threshold_cell_lines <- "10percent"

#1. creating plots for absolute correlations
barchart_absolute <- NULL

#merge results for each cancer type
for (cancer in cancer_list) {
  if(cancer == "pancancer") {
    barchart_absolute_cancer <- read.csv(paste0("./results/barcharts_data/", threshold_cell_lines, "_",
                                                cancer, "_", regulons, "_number_signif_corr.csv")) %>%
      mutate_at("Cancer", .funs = toupper)     
  } else {
    barchart_absolute_cancer <- read.csv(paste0("./results/barcharts_data/", 
                                                cancer, "_", regulons, "_number_signif_corr.csv")) %>%
      mutate_at("Cancer", .funs = toupper)  
  }

  barchart_absolute <- rbind(barchart_absolute, barchart_absolute_cancer)  
}

#add number of genes to each cancer as a string so it can be in each plot's header
barchart_absolute$Cancer <- str_c(barchart_absolute$Cancer,
                         '   |   No. genes = ', barchart_absolute$total_drivers)

#sorting by best performing method
vector_sorted_best_method <- barchart_absolute %>%
  group_by(Method) %>%
  summarise(sum(Percentage_signif)) %>%
  set_colnames(c("Method", "sum_percentage_signif")) %>%
  arrange(sum_percentage_signif, decreasing = TRUE) %>%
  pull(Method) %>%
  as.vector()

#make factors for ggplot to arrange in order of best to worst performing method
barchart_absolute$Method <- factor(barchart_absolute$Method, 
                          levels = vector_sorted_best_method)

#sets parameters for plot according to type of analysis chosen: either pancancer or per cancer type
if (cancer == "pancancer") {
  height <- 7.5
  width <- 15
  filename_absolute <- paste0("./plots/barcharts/", threshold_cell_lines, "_", regulons, "_barcharts_5methods.pdf")
  filename_pos_vs_neg <- paste0("./plots/barcharts/", threshold_cell_lines, "_", regulons, "_pos_vs_neg_5methods.pdf")
  axis_label_absolute <- "Genes significantly correlated\nwith CRISPR gene sensitivity (%)"
  axis_label_pos_vs_neg <- "Genes significantly correlated\nwith CRISPR gene sensitivity (count)"
} else {
  height <- 10
  width <- 38
  filename_absolute <- paste0("./plots/barcharts/", regulons, "_barcharts_5methods.pdf")
  filename_pos_vs_neg <- paste0("./plots/barcharts/", regulons, "_pos_vs_neg_5methods.pdf")
  axis_label_absolute <- "Genes significantly correlated with CRISPR gene sensitivity (%)"
  axis_label_pos_vs_neg <- "Genes significantly correlated with CRISPR gene sensitivity (count)"
}
width <- 20
#faceted barchart absolute correlations
plot_barchart_absolute <- barchart_absolute %>%
  ggplot() +
  geom_bar(aes(fill = Pearsons_R, x = Method, y = Percentage_signif),
           stat = "identity", position = position_stack(reverse = FALSE)) +
  coord_flip() +
  scale_fill_manual(values = c("#A9D6E5", "#61A5C2", "#2C7DA0", "#01497C"), 
                    name = "Pearson's R") +
  theme_bw(base_size = 16) +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        panel.background = element_blank()) +
  geom_text(x = 25, y = 1, label = barchart_absolute$total_drivers) +
  labs(title = paste0(regulons, " regulons"), x = "Method", y = axis_label_absolute) +
  facet_wrap(vars(Cancer), nrow = 2)  


#2. creating plots for pos vs neg correlations
barchart_pos_vs_neg <- NULL

for (cancer in cancer_list) {
  if(cancer == "pancancer") {
    barchart_pos_vs_neg_cancer <- read.csv(paste0("./results/barcharts_data/", threshold_cell_lines, "_",
                                                  cancer, "_", regulons, "_pos_vs_neg_corr.csv")) %>%
      mutate_at("Cancer", .funs = toupper)        
  } else {
    barchart_pos_vs_neg_cancer <- read.csv(paste0("./results/barcharts_data/", 
                                                  cancer, "_", regulons, "_pos_vs_neg_corr.csv")) %>%
      mutate_at("Cancer", .funs = toupper)     
  }

  barchart_pos_vs_neg <- rbind(barchart_pos_vs_neg, barchart_pos_vs_neg_cancer)  
}

barchart_pos_vs_neg$count <- ifelse(barchart_pos_vs_neg$Pearsons_R == "R≤0", -1 * barchart_pos_vs_neg$count, barchart_pos_vs_neg$count)

barchart_pos_vs_neg$Method <- factor(barchart_pos_vs_neg$Method, 
                          levels = vector_sorted_best_method)

#faceted barchart pos vs neg correlations
plot_barchart_pos_vs_neg <-
  barchart_pos_vs_neg %>%
  ggplot() +
  geom_bar(aes(x = Method, y = count, fill = Pearsons_R), stat = "identity", 
           position = position_stack(reverse = FALSE)) +
  scale_fill_manual(values = c("#F7B05B", "#C2E1C2"), 
                    name = "Pearson's R", 
                    labels=c(expression("\u2191 expression\n\u2191 sensitivity"), 
                             expression("\u2191 expression\n\u2193 sensitivity"))) +
  geom_text(data = subset(barchart_pos_vs_neg, count != 0), 
            aes(x = Method, y = count, label = abs(count)), 
            hjust = ifelse(subset(barchart_pos_vs_neg, count != 0)$count > 0, 1, -0.1)) +
  scale_y_continuous(labels = abs) +
  coord_flip() +
  theme_bw(base_size = 16) + 
  theme(panel.grid = element_blank(),
        legend.text.align = 0) +
  labs(title = paste0(regulons, " regulons"), x = "Method", y = axis_label_pos_vs_neg) +
  facet_wrap(vars(Cancer), nrow = 2)  

#saving plots
ggsave(filename = filename_absolute, plot = plot_barchart_absolute, 
       width = width, height = height, 
       dpi = 1000, units = "cm", device = cairo_pdf) 

ggsave(filename = filename_pos_vs_neg, plot = plot_barchart_pos_vs_neg, 
       width = width, height = height, 
       dpi = 1000, units = "cm", device = cairo_pdf)
